refactor: extract canonical PDF source anchors
